# Special case as_draws_array for length 1