When we use as_draws_array and we only have 1 chain we could optimize the function a bit. 

Testing scenario: 
```
library(cmdstanr)
model_code <- "
parameters {
  real k;
}

transformed parameters {
  real x[50000] = rep_array(k, 50000);
}

model {
  k ~ normal(0, 1);
}
"
model_file <- write_stan_file(model_code)
mod <- cmdstan_model(model_file)
fit <- mod$sample(iter_sampling = 2000, validate_csv = FALSE, chains = 1)

data <- data.table::fread(cmd= paste0("grep -v '^#' ", fit$output_files()[1]))

for (i in 1:10) {
  start <- Sys.time()
  d_a <- posterior::as_draws_array(data)
  print(Sys.time()-start)
}
```

this branch:
```
Time difference of 2.206097 secs
Time difference of 1.762325 secs
Time difference of 1.766448 secs
Time difference of 1.793118 secs
Time difference of 1.778449 secs
Time difference of 1.747509 secs
Time difference of 1.872519 secs
Time difference of 1.80685 secs
Time difference of 1.922842 secs
Time difference of 1.815559 secs
```

master:
```
Time difference of 3.483295 secs
Time difference of 2.871998 secs
Time difference of 2.778586 secs
Time difference of 2.923501 secs
Time difference of 2.722884 secs
Time difference of 2.784933 secs
Time difference of 2.740157 secs
Time difference of 2.707214 secs
Time difference of 2.70872 secs
Time difference of 2.746846 secs
```

I am looking into other bottlenecks so if you prefer to review in larger chunks we can keep the PR open and I can note when I am done.