@paul-buerkner 2020-11-14
Thank you! I am fine with both, keeping the PR open for more changes or merging just this fix. What is your preference?