@jgabry 2020-11-16
Thanks! It looks good to me too. Should we bump the version number when we merge this? That way we can require this version for cmdstanr.