@rok-cesnovar 2020-11-16
Lets hold off for a few more days with the bump, I might have time to look into other bottlnecks causing #98 
If I dont find any till next week, we can then bump.