@jgabry 2020-11-16
Since @paul-buerkner already approve this I'll go ahead and merge now and then we can do separate PRs for other fixes (e.g., if you or someone else figures out #98)